Wedding Text Analysis NLP and Sentiment Insights/requirements.txt
newspaper3k
pandas
nltk
textblob
matplotlib
seaborn
wordcloud